Using INT_MAX instead of ULONG_MAX to avoid Xerces issue
[#117406567]

+++ libgpopt/include/gpopt/engine/CHint.h
@@ -69,7 +69,7 @@ namespace gpopt
 			static
 			CHint *PhintDefault(IMemoryPool *pmp)
 			{
-				return GPOS_NEW(pmp) CHint(ULONG_MAX /* ulMinNumOfPartsToRequireSortOnInsert */);
+				return GPOS_NEW(pmp) CHint(INT_MAX /* ulMinNumOfPartsToRequireSortOnInsert */);
 			}
 
 	}; // class CHint
